# swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /categorys:
    get:
      summary: Get a list of all categorys
      description: Restricted to admin users.
      operationId: getAllCategorys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of all categorys
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    name:
                      type: string
                      example: "Merah"
                    type:
                      type: string
                      example: "main"
                    position:
                      type: number
                      example: 1
                    isActive:
                      type: boolean
                      example: true
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new category
      description: Restricted to admin users.
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Merah"
                type:
                  type: string
                  enum: [Admin, User]
                position:
                  type: number
                  example: 1
                isActive:
                  type: boolean
                  example: true
              required:
                - name
                - type
                - position
                - isActive
      responses:
        "200":
          description: Category created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  name:
                    type: string
                    example: "Merah"
                  type:
                    type: string
                    example: "main"
                  position:
                    type: number
                    example: 1
                  isActive:
                    type: boolean
                    example: true
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "400":
          description: Name is already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Name 'Merah' is already registered"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /categorys/{id}:
    parameters:
      - in: path
        name: id
        description: Category id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single category by id
      description: Admin users can access any account, regular users are restricted to their own account.
      operationId: getCategoryById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified category
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  name:
                    type: string
                    example: "Merah"
                  type:
                    type: string
                    example: "main"
                  position:
                    type: number
                    example: 1
                  isActive:
                    type: boolean
                    example: true
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update category
      description: Restricted to admin users.
      operationId: updateCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Merah"
                type:
                  type: string
                  enum: [Admin, User]
                position:
                  type: number
                  example: 1
                isActive:
                  type: boolean
                  example: true
              required:
                - name
                - type
                - position
                - isActive
      responses:
        "200":
          description: Category created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  name:
                    type: string
                    example: "Merah"
                  type:
                    type: string
                    example: "main"
                  position:
                    type: number
                    example: 1
                  isActive:
                    type: boolean
                    example: true
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "400":
          description: Name is already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Name 'Merah' is already registered"
        "401":
          $ref: "#/components/responses/UnauthorizedError"   
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"